Small bug fixes
Overlooked launch scenario for console launch.
Mistakenly stored int parameter as double that could result in crash.

+++ src/idynomics/Global.java
@@ -305,7 +305,7 @@ public class Global extends ParameterSet
 	/* 
 	 * Default maximum displacement per step, set default if none.
 	 */
-	public static double mechanical_max_iterations = 10000;
+	public static Integer mechanical_max_iterations = 10000;
 	
 	/*
 	 * Default mechanical stress threshold at which a system may be considered
